Update CDX vulnerability and related types to v1.5
