+++ cmd/vulnerability.go
@@ -348,7 +348,7 @@ func hashVulnerability(cdxVulnerability schema.CDXVulnerability, whereFilters []
 
 	// hash any component w/o a license using special key name
 	vulnInfo.Vulnerability = cdxVulnerability
-	vulnInfo.BomRef = cdxVulnerability.BomRef.String()
+	vulnInfo.BomRef = cdxVulnerability.BOMRef.String()
 	vulnInfo.Id = cdxVulnerability.Id
 
 	// Truncate dates from 2023-02-02T00:00:00.000Z to 2023-02-02
