+++ schema/cyclonedx_common.go
@@ -17,11 +17,11 @@
 
 package schema
 
-// v1.4 added
-// v1.5 added Constraints: "minLength": 1
+// v1.4: added
+// v1.5: added Constraints: "minLength": 1
 type CDXRefType string
 
-// v1.5 added Stringer interface
+// v1.5: added Stringer interface
 func (ref CDXRefType) String() string {
 	return string(ref)
 }
@@ -115,4 +115,24 @@ type CDXProperty struct {
 type CDXLicenseLegalParty struct {
 	Organization CDXOrganizationalEntity  `json:"organization,omitempty"`
 	Individual   CDXOrganizationalContact `json:"individual,omitempty"`
+}
+
+// v1.2: existed
+// v1.4: added "externalReferences"
+// v1.5: deprecated "Creation Tools (legacy)" object in favor of new "Creation Tools" object
+// - v1.5 Note: The v1.4 structure/fields is now called the "Creation Tools (legacy)" structure
+// - v1.5: In order to support the new object "Creation Tools", we need to combine these fields
+// into with the legacy structure fields
+type CDXLegacyCreationTool struct {
+	Vendor             string                 `json:"vendor,omitempty" cdx:"deprecated"`       // v1.5: deprecated
+	Name               string                 `json:"name,omitempty" cdx:"deprecated"`         // v1.5: deprecated
+	Version            string                 `json:"version,omitempty" cdx:"deprecated"`      // v1.5: deprecated
+	Hashes             []CDXHash              `json:"hashes,omitempty" cdx:"deprecated"`       // v1.5: deprecated
+	ExternalReferences []CDXExternalReference `json:"externalReferences,omitempty" cdx:"+1.4"` // v1.4: added, v1.5: deprecated
+}
+
+// v1.5: created. Intended to be used instead of (legacy) Creation Tools which was deprecated
+type CDXCreationTools struct {
+	Components []CDXComponent `json:"components,omitempty" cdx:"+1.5"` // v1.5: added (new type)
+	Services   []CDXService   `json:"services,omitempty" cdx:"+1.5"`   // v1.5: added (new type)
 }
